Revert "Fixes #740: optimized chunks size"
