+++ package.json
@@ -9,11 +9,11 @@
     "babel-eslint": "4.1.8",
     "babel-loader": "6.2.4",
     "babel-plugin-add-module-exports": "0.1.4",
-    "babel-plugin-object-assign": "1.2.1",
-    "babel-plugin-react-transform": "2.0.2",
     "babel-preset-es2015": "6.6.0",
     "babel-preset-react": "6.5.0",
     "babel-preset-stage-0": "6.5.0",
+    "babel-plugin-object-assign": "1.2.1",
+    "babel-plugin-react-transform": "2.0.2",
     "css-loader": "0.19.0",
     "download-cli": "1.0.1",
     "escope": "3.2.0",
@@ -43,8 +43,8 @@
     "mocha": "2.4.5",
     "ncp": "2.0.0",
     "parallelshell": "1.2.0",
-    "react-addons-css-transition-group": "0.14.8",
     "react-addons-test-utils": "0.14.8",
+    "react-addons-css-transition-group": "0.14.8",
     "react-hot-loader": "1.3.0",
     "react-router": "2.4.0",
     "react-router-redux": "2.1.0",
@@ -59,7 +59,6 @@
     "style-loader": "0.12.4",
     "url-loader": "0.5.7",
     "webpack": "1.12.15",
-    "webpack-bundle-size-analyzer": "2.0.2",
     "webpack-dev-server": "1.14.1"
   },
   "//": "replace react-sortable-items with official on npm when 0.0.10 with remove_deprecated PR in included",
@@ -129,7 +128,6 @@
     "postinstall": "node checkCesium.js && mkdirp ./web/client/libs/Cesium && download --extract --out ./web/client/libs/Cesium https://cesiumjs.org/releases/Cesium-1.17.zip || echo Cesium already installed",
     "clean": "rimraf ./web/client/dist",
     "compile": "npm run clean && mkdirp ./web/client/dist && webpack --config prod-webpack.config.js",
-    "analyze": "npm run clean && mkdirp ./web/client/dist && webpack --json | webpack-bundle-size-analyzer",
     "start": "webpack-dev-server --progress --colors --port 8081 --hot --inline --content-base web/client",
     "examples": "webpack-dev-server --progress --colors --port 8081 --hot --inline --content-base web/client --config examples-webpack.config.js",
     "test": "karma start ./karma.conf.single-run.js",
