+++ web/client/examples/rasterstyler/pages/MapViewer.jsx
@@ -23,8 +23,7 @@ const urlQuery = url.parse(window.location.href, true).query;
 
 const PluginsContainer = connect((state) => ({
     pluginsConfig: state.plugins || ConfigUtils.getConfigProp('plugins') || null,
-    mode: (urlQuery.mobile || (state.browser && state.browser.touch)) ? 'mobile' : 'desktop',
-    pluginsState: state
+    mode: (urlQuery.mobile || (state.browser && state.browser.touch)) ? 'mobile' : 'desktop'
 }))(require('../../../components/plugins/PluginsContainer'));
 
 
@@ -66,3 +65,4 @@ module.exports = connect((state) => ({
     loadMapConfig,
     reset: resetControls
 })(MapViewer);
+
