+++ web/client/plugins/BurgerMenu.jsx
@@ -56,8 +56,7 @@ const BurgerMenu = React.createClass({
         };
     },
     getPanels() {
-        return this.props.items.filter((item) => item.panel)
-            .map((item) => assign({}, item, {panel: item.panel === true ? item.plugin : item.panel}));
+        return this.props.items.filter((item) => item.panel);
     },
     getTools() {
         return [{element: <span key="burger-menu-title">{this.props.title}</span>}, ...this.props.items.sort((a, b) => a.position - b.position)];
